feat: add create-cli-utils logic
