+++ packages/build-scripts/src/index.ts
@@ -1,11 +1,16 @@
-import build = require('./commands/build')
-import start = require('./commands/start')
-import test = require('./commands/test')
+import start = require('./apis/start');
+import build = require('./apis/build');
+import test = require('./apis/test');
+import childProcessStart = require('./apis/child-process-start');
+import createCli = require('./apis/create-cli');
 
 export * from './core/Context';
 export * from './types';
+
 export {
-  build,
   start,
+  build,
   test,
+  childProcessStart,
+  createCli,
 };
