+++ packages/build-scripts/src/utils/checkNodeVersion.ts
@@ -1,10 +1,10 @@
 import semver = require('semver')
 import log = require('./log')
 
-export = function checkNodeVersion(requireNodeVersion: string): void {
+export = function checkNodeVersion(requireNodeVersion: string, frameworkName = 'build-scripts'): void {
   if (!semver.satisfies(process.version, requireNodeVersion)) {
     log.error('ENV', `You are using Node ${process.version}`);
-    log.error('ENV', `build-scripts requires Node ${requireNodeVersion}, please update Node.`);
+    log.error('ENV', `${frameworkName} requires Node ${requireNodeVersion}, please update Node.`);
     process.exit(1);
   }
 };
